pyproject.toml
[tool.poetry]
name = "siim-isic-melanoma-classification"
version = "0.1.0"
description = ""
authors = ["MasatakaImamura <masataka.i.navi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.5"
matplotlib = "^3.2.2"
seaborn = "^0.10.1"
jupyter = "^1.0.0"
scikit-learn = "^0.23.1"
tqdm = "^4.47.0"
kaggle = "^1.5.6"
mlflow = "^1.9.1"
hydra-core = "^0.11.3"
pydicom = "^2.0.0"
jupyterlab = "^2.1.5"
pandas-profiling = "^2.8.0"
opencv-python = "^4.3.0"
albumentations = "^0.4.6"
comet_ml = "^3.1.12"
torch_optimizer = "^0.0.1-alpha.13"
efficientnet-pytorch = "^0.6.3"
category_encoders = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
